+++ concordium-consensus/src/Concordium/GlobalState/BlockState.hs
@@ -203,6 +203,9 @@ class AccountOperations m => BlockStateQuery m where
     -- |Query an account by the id of the credential that belonged to it.
     getAccountByCredId :: BlockState m -> CredentialRegistrationID -> m (Maybe (AccountIndex, Account m))
 
+    -- |Query an account by the account index that belonged to it.
+    getAccountByIndex :: BlockState m -> AccountIndex -> m (Maybe (AccountIndex, Account m))
+
     -- |Get the contract state from the contract table of the state instance.
     getContractInstance :: BlockState m -> ContractAddress -> m (Maybe Instance)
 
@@ -644,6 +647,7 @@ instance (Monad (t m), MonadTrans t, BlockStateQuery m) => BlockStateQuery (MGST
   getAccount s = lift . getAccount s
   accountExists s = lift . accountExists s
   getAccountByCredId s = lift . getAccountByCredId s
+  getAccountByIndex s = lift . getAccountByIndex s
   getBakerAccount s = lift . getBakerAccount s
   getContractInstance s = lift . getContractInstance s
   getModuleList = lift . getModuleList
@@ -674,6 +678,7 @@ instance (Monad (t m), MonadTrans t, BlockStateQuery m) => BlockStateQuery (MGST
   {-# INLINE getAccount #-}
   {-# INLINE accountExists #-}
   {-# INLINE getAccountByCredId #-}
+  {-# INLINE getAccountByIndex #-}
   {-# INLINE getBakerAccount #-}
   {-# INLINE getContractInstance #-}
   {-# INLINE getModuleList #-}